Moved filename extension filtering into fs-migrator
