+++ src/migrate/sources/fs-migrations.js
@@ -1,25 +1,37 @@
 import fs from 'fs';
 import path from 'path';
 import Promise from 'bluebird';
-import { sortBy } from 'lodash';
+import { sortBy, filter } from 'lodash';
 
 const readDirAsync = Promise.promisify(fs.readdir, { context: fs });
 
+export const DEFAULT_LOAD_EXTENSIONS = Object.freeze([
+  '.co',
+  '.coffee',
+  '.eg',
+  '.iced',
+  '.js',
+  '.litcoffee',
+  '.ls',
+  '.ts',
+]);
+
 export default class FsMigrations {
-  constructor(migrationDirectories, sortDirsSeparately) {
+  constructor(migrationDirectories, sortDirsSeparately, loadExtensions) {
     this.sortDirsSeparately = sortDirsSeparately;
 
     if (!Array.isArray(migrationDirectories)) {
       migrationDirectories = [migrationDirectories];
     }
     this.migrationsPaths = migrationDirectories;
+    this.loadExtensions = loadExtensions || DEFAULT_LOAD_EXTENSIONS;
   }
 
   /**
    * Gets the migration names
    * @returns Promise<string[]>
    */
-  getMigrations() {
+  getMigrations(loadExtensions) {
     // Get a list of files in all specified migration directories
     const readMigrationsPromises = this.migrationsPaths.map((configDir) => {
       const absoluteDir = path.resolve(process.cwd(), configDir);
@@ -47,10 +59,18 @@ export default class FsMigrations {
       // If true we have already sorted the migrations inside the folders
       // return the migrations fully qualified
       if (this.sortDirsSeparately) {
-        return migrations;
+        return filterMigrations(
+          this,
+          migrations,
+          loadExtensions || this.loadExtensions
+        );
       }
 
-      return sortBy(migrations, 'file');
+      return filterMigrations(
+        this,
+        sortBy(migrations, 'file'),
+        loadExtensions || this.loadExtensions
+      );
     });
   }
 
@@ -62,4 +82,12 @@ export default class FsMigrations {
     const absoluteDir = path.resolve(process.cwd(), migration.directory);
     return require(path.join(absoluteDir, migration.file));
   }
+}
+
+function filterMigrations(migrationSource, migrations, loadExtensions) {
+  return filter(migrations, (migration) => {
+    const migrationName = migrationSource.getMigrationName(migration);
+    const extension = path.extname(migrationName);
+    return loadExtensions.includes(extension);
+  });
 }
